refactor: unify code to avoid duplication raised by sonar
