+++ apps/ledger-live-mobile/src/components/WebPlatformPlayer/CustomHandlers.ts
@@ -16,6 +16,7 @@ import { addOneAccount } from "~/actions/accounts";
 import { setAccountName } from "@ledgerhq/live-wallet/store";
 import { handlers as deeplinkHandlers } from "@ledgerhq/live-common/wallet-api/CustomDeeplink/server";
 import { handlers as liveAppModalHandlers } from "@ledgerhq/live-common/wallet-api/LiveAppModal/server";
+import { resolveLiveAppModalParams } from "@ledgerhq/live-common/wallet-api/LiveAppModal/types";
 import { setLiveAppModal } from "~/reducers/liveAppModal";
 import { Linking } from "react-native";
 
@@ -140,26 +141,8 @@ export function useLiveAppModalCustomHandlers(manifest: WebviewProps["manifest"]
     return {
       ...liveAppModalHandlers({
         uiHooks: {
-          "custom.liveApp.modal.open": ({
-            requestId,
-            manifestId,
-            path,
-            title,
-            description,
-            size,
-            useCase,
-          }) => {
-            dispatch(
-              setLiveAppModal({
-                requestId,
-                manifestId: manifestId || manifest.id,
-                path,
-                title,
-                description,
-                size,
-                useCase,
-              }),
-            );
+          "custom.liveApp.modal.open": input => {
+            dispatch(setLiveAppModal(resolveLiveAppModalParams(input, manifest.id)));
             navigation.navigate(ScreenName.LiveAppModal);
           },
         },
