+++ apps/ledger-live-mobile/src/reducers/liveAppModal.ts
@@ -1,32 +1,12 @@
-import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
-import type {
-  LiveAppModalSize,
-  LiveAppModalUseCase,
-} from "@ledgerhq/live-common/wallet-api/LiveAppModal/types";
 import type { State } from "~/reducers/types";
 
-export type LiveAppModalParams = {
-  requestId: string;
-  manifestId: string;
-  path: string;
-  title?: string;
-  description?: string;
-  size?: LiveAppModalSize;
-  useCase?: LiveAppModalUseCase;
-};
-
-export type LiveAppModalState = LiveAppModalParams | null;
-
-const liveAppModalSlice = createSlice({
-  name: "liveAppModal",
-  initialState: null as LiveAppModalState,
-  reducers: {
-    setLiveAppModal: (_state, action: PayloadAction<LiveAppModalParams | null>) => action.payload,
-  },
-});
-
-export const { setLiveAppModal } = liveAppModalSlice.actions;
+export type {
+  LiveAppModalParams,
+  LiveAppModalState,
+} from "@ledgerhq/live-common/wallet-api/LiveAppModal/types";
+export {
+  setLiveAppModal,
+  liveAppModalReducer as default,
+} from "@ledgerhq/live-common/wallet-api/LiveAppModal/slice";
 
 export const selectLiveAppModal = (state: Pick<State, "liveAppModal">) => state.liveAppModal;
-
-export default liveAppModalSlice.reducer;
